// src/types/ui.ts
export interface UIElements {
  volumeSlider: HTMLInputElement;
  volumeText: HTMLInputElement;
  muteCheckbox: HTMLInputElement;
  displayModeCheckBox: HTMLInputElement;
  undoButton: HTMLButtonElement;
  tabList: HTMLUListElement;
  popupContent: HTMLElement;
  errorContent: HTMLElement;
  body: HTMLBodyElement;
}